requirements.txt
fastapi==0.118.0
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf==6.32.1
typing_extensions==4.15.0
